refactor(crew-ai): move the dojo into its own examples project
Every other integration with a dev server keeps it in a `python/examples/`
project next to the library, declares `dev` there, and publishes a library
that ships no scripts. crew-ai was the only one that kept its dojo server
and demo flows inside the publishable package, which is why a library ended
up advertising a `dev` console script in the first place.

The dojo server and the twenty demo flows move to
`integrations/crew-ai/python/examples/agents/`, a separate project with its
own pyproject, `dev = "agents.dojo:main"`, and a path dependency on the
library so the server still exercises the bridge in this checkout rather
than a release. Git recorded the whole move as renames, so per-file history
survives.

Imports: the demos reached into the library through `..sdk`, `..a2ui_tool`
and `.._hitl`, which become absolute `ag_ui_crewai.*`. Imports between demos
were already relative and move unchanged.

The telemetry opt-out that makes Ctrl-C work could not come along inside
`dojo.py`. `ag_ui_crewai/__init__.py` imports crewai, and `dojo.py` imports
the bridge at module scope, so by the time its body runs the SIGINT handler
is already installed. It lives in `agents/__init__.py` instead, which runs
first when the entry point imports `agents.dojo`. Verified: after
`agents/__init__` the variable is set and crewai is not yet loaded.

Verified: `uv sync` and `uv run dev` in the new project serve all 28 routes
including the 13 conversational ones, and SIGINT to the server's own process
group exits it in 1.5s. The hang the opt-out prevents needs queued spans, so
starting and stopping a server without driving a flow does not reproduce it;
what is verified here is the ordering property the fix depends on, not the
fix end to end.

Also declares what the demos import directly (crewai, litellm, pydantic,
fastapi, the a2ui toolkit) rather than relying on the library's transitives.
