+++ integrations/crew-ai/python/examples/agents/_model_turn.py
