+++ integrations/crew-ai/python/examples/agents/a2ui_fixed_schema_schemas/flight_schema.json
