+++ integrations/crew-ai/python/examples/agents/agentic_chat.py
@@ -4,7 +4,7 @@ A simple agentic chat flow.
 
 from crewai.flow.flow import Flow, start
 from litellm import acompletion
-from ..sdk import copilotkit_stream, CopilotKitState
+from ag_ui_crewai.sdk import copilotkit_stream, CopilotKitState
 
 class AgenticChatFlow(Flow[CopilotKitState]):
