+++ integrations/crew-ai/python/examples/agents/human_in_the_loop.py
@@ -6,7 +6,7 @@ from crewai.flow.flow import Flow, start, router, listen
 from litellm import acompletion
 from pydantic import BaseModel
 from typing import Literal, List
-from ..sdk import (
+from ag_ui_crewai.sdk import (
   copilotkit_stream,
   CopilotKitState,
 )
